feat: Make IntelliSense function
Not perfect it takes a long time for the database to be
reset when we swtich to different sketches and it has
to be rebuilt each time.

The has some suggestions, see `Vscode LSP setup` chat.
If/when I "solve" this I'll update README.md and point
the the chat.